% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced.R
\name{balance.miss}
\alias{balance.miss}
\title{Function to search for the indicator planned missing designs using Mplus}
\usage{
balance.miss(VNAMES, distal.var = NULL, n, nreps, seed, Time, k,
  Time.complete, costmx, pc, pd, design0.out, focal.param,
  eval.budget = T, rm.budget = NULL, complete.var = NULL)
}
\arguments{
\item{VNAMES}{The names of the observed variables, ordered by the time they are measured}

\item{distal.var}{Any distal variables included in the model that would have complete data}

\item{n}{The total sample size as initially planned}

\item{nreps}{Number of replications for Monte Carlo simulation for each possible design}

\item{seed}{Random seet for simulation}

\item{Time}{The total number of time points (waves of data collection)}

\item{k}{The number of observed variables collected at each wave}

\item{Time.complete}{Number of waves of data collection that have been completed before the funding cut occured}

\item{costmx}{The vector containing the unit cost of each observed variable which has no data collected yet. They are constant across subjects, but they can vary across variables and across time.}

\item{pc}{The proportion of subjects that will participate in all of the following waves of data collection and provide complete data (must be greater than 0)}

\item{pd}{The proportion of subjects that will not participate in any of the following waves of data collection (i.e., attritors). This can be 0.}

\item{design0.out}{Mplus output file which contains the a priori power analysis/sample size planning (simulation) results for this
specific model assuming a complete data design. Theoretically, such analysis was supposed to be conducted before the study began.}

\item{focal.param}{The parameters of focal interest. If engine="l", the focal parameters should be specified using
the lavaan script. If engine="m", the focal parameters should be specified based on the Mplus output file design0.out.}

\item{eval.budget}{Budget constraints. If the researcher wishes to search for designs under the budget limit, you to provide the remaining available budget that can be used for data collection}

\item{rm.budget}{The amount of remaining budget avaialbe for data collection}

\item{complete.var}{Specify if there are any variables that need to have complete data collected across all the participating subjects}
}
\value{
An object containing the information of the optimal PM design, with highest power for testing the
focal parameters, compared with other PM designs
}
\description{
Function to search for the indicator planned missing designs using Mplus
}
\seealso{
\code{\link{simPM}} which wraps this function
}
