% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpattern.R
\name{plotPM}
\alias{plotPM}
\title{Plot the missing data patterns for the optimal PM design.}
\usage{
plotPM(object, colbr = "PRGn", col = c("antiquewhite1", "firebrick"),
  row.names = T, labels = T, fontsize_col = 20, fontsize_row = 14,
  fontsize = 14, angle_col = 45, legend = T, main = "")
}
\arguments{
\item{object}{The simPM object}

\item{colbr}{colors for waves. Default is "PRGn".}

\item{col}{colors for complete vs. missing data. Default is c("antiquewhite1","firebrick").}

\item{labels}{logical, indicating whether the label for waves is needed. Default is T.}

\item{fontsize_col}{specify the font size for the column labels. Default is 20.}

\item{fontsize_row}{specify the font size for the row labels. Default is 14.}

\item{fontsize}{specify the font size for the legend. Default is 14.}

\item{angle_col}{specify the angle of how the column labels are displayed}

\item{legend}{logical, indicating whether the legend is shown. Default is T.}

\item{main}{specify the plot title}
}
\description{
Plot the missing data patterns for the optimal PM design.
}
\examples{
\dontrun{
plotPM(wave.out)
plotPM(indicator.out)
plotPM(forward.out,labels=F,col=c("gray96","gray35"),fontsize_row=26,fontsize=18,fontsize_col=26)
}
}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
