% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres-cost.R
\name{pres.cost}
\alias{pres.cost}
\title{To examine the lower level designs in forward selection}
\usage{
pres.cost(opt.pattern, costmx, max.mk, pc, pd, n, k, Time, Time.complete)
}
\value{
An object containing the information of the optimal PM design, with highest power for testing the
focal parameters, compared with other PM designs
}
\description{
To examine the lower level designs in forward selection
}
\seealso{
\code{\link{simPM}} which wraps this function
}
