% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres-design.R
\name{pres.design}
\alias{pres.design}
\title{Check the simulation results for lower level designs in forward assembly. 
\code{pres.design} is only applicable when Mplus is used for simulations.}
\usage{
pres.design(opt.results, max.mk)
}
\arguments{
\item{opt.results}{The object returned from forward assembly.}

\item{max.mk}{The maximum number of unique missing data
patterns in the selected design.}
}
\description{
Check the simulation results for lower level designs in forward assembly. 
\code{pres.design} is only applicable when Mplus is used for simulations.
}
