% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.opt}
\alias{summary.opt}
\title{A summary function to extract the important information of the output object.}
\usage{
summary.opt(object)
}
\value{
An object containing the information of the optimal PM design, with highest power for testing the
focal parameters, compared with other PM designs
}
\description{
A summary function to extract the important information of the output object.
}
