% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.simpm}
\alias{summary.simpm}
\title{Extract the summary information of the simpm object.}
\usage{
\method{summary}{simpm}(object)
}
\arguments{
\item{object}{The output object returned by \code{\link{simPM}}.}
}
\value{
The information about the optimal PM design.
}
\description{
\code{summary.simpm} summarizes and extracts the important information about the optimal PM design.
}
