% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave-missing.R
\name{wave.miss}
\alias{wave.miss}
\title{Search for the optimal wave-level PM designs.}
\usage{
wave.miss(VNAMES, distal.var = NULL, n, nreps, seed, Time, k,
  Time.complete, costmx, pc, pd, design0.out, focal.param,
  eval.budget = T, rm.budget = NULL, complete.wave = NULL)
}
\arguments{
\item{VNAMES}{A character vector containing the names of the observed
variables. The variable names must be ordered chronologically, 
by the time (wave) they are measured.}

\item{distal.var}{Char vector. Specify the names of the variables, if
there are any time-independent distal variables included in the
model that are not subject to planned missingness.}

\item{n}{The total sample size as initially planned.}

\item{nreps}{Number of replications for Monte Carlo simulations.}

\item{seed}{The seed for random number generation.}

\item{Time}{The total number of time points (i.e., the total number 
of data collection waves).}

\item{k}{The number of observed variables collected at each wave.}

\item{Time.complete}{Number of data collection waves that have been
completed before the funding cut occurs.}

\item{costmx}{A numeric vector containing the unit cost of each
observed variable that is yet to be measured (post the funding cut).
The cost is assumed to be constant across subjects, but it is
allowed to vary across variables and across waves.}

\item{pc}{Numeric. Proportion of completers: the proportion of subjects
that will participate in all of the following waves of data
collection and provide complete data. This must be greater than 0.}

\item{pd}{Numeric. The proportion of subjects that will not participate
in any of the following waves of data collection (i.e., drop from
the longitudinal study). This value can be 0.}

\item{design0.out}{An object returned by \code{\link[MplusAutomation]{r#'    eadModels}}. To obtain this object, the user need to have a Mplus 
output file which contains the \emph{a priori} power analysis 
results for this specific model assuming a complete data design 
(i.e., simulation-based power analysis for sample size planning). 
In principle, \emph{a priori} power analysis is supposed to be 
conducted before the study began.}

\item{focal.param}{Char vector. The parameters of focal interest. The
focal parameters should be specified in the specific format based on
the Mplus output object \code{design0.out}.}

\item{eval.budget}{Logical scalar (TRUE or FALSE), indicating whether
there is any budget constraint. If the user wishes to search for PM
designs under the budget limit, they need to specify the amount of
the remaining available budget that can be used for future data
collection.}

\item{rm.budget}{Numeric. The amount of remaining budget avaialbe for
future data collection.}

\item{complete.wave}{Numeric vector. Specify which wave(s) that the
user wish to have complete data collected from all the participants.}
}
\value{
An object containing the information of the optimal
   wave-level missing design. The optimal design is the one that yields
   highest power for testing the focal parameters, compared to other
   plausible candidate PM designs.
}
\description{
\code{wave.miss} runs simulations using M\emph{plus}. It returns the search results for optimal wave-level PM designs.
}
\seealso{
\code{\link{simPM}} which is a warpper function for this
   function.
}
