% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced-simsem.R
\name{balance.miss.l}
\alias{balance.miss.l}
\title{Function to search for the optimal planned missing design using lavaan/simsem}
\usage{
balance.miss.l(popModel, analyzeModel, NAMES, Time, Time.complete, k, pc,
  pd, costmx, n, nreps, focal.param, complete.var = NULL,
  eval.budget = T, rm.budget = NULL, distal.var = NULL,
  seed = 1234)
}
\arguments{
\item{popModel}{The data generation model (population model) specified using lavaan script}

\item{analyzeModel}{The analysis model, which can be different from the population model, using lavaan script specific model assuming a complete data design. Theoretically, such analysis was supposed to be conducted before the study began.}

\item{NAMES}{The names of the observed variables, ordered by the time they are measured}

\item{Time}{The total number of time points (waves of data collection)}

\item{Time.complete}{Number of waves of data collection that have been completed before the funding cut occured}

\item{k}{The number of observed variables collected at each wave}

\item{pc}{The proportion of subjects that will participate in all of the following waves of data collection and provide complete data (must be greater than 0)}

\item{pd}{The proportion of subjects that will not participate in any of the following waves of data collection (i.e., attritors). This can be 0.}

\item{costmx}{The vector containing the unit cost of each observed variable which has no data collected yet. They are constant across subjects, but they can vary across variables and across time.}

\item{n}{The total sample size as initially planned}

\item{nreps}{Number of replications for Monte Carlo simulation for each possible design}

\item{focal.param}{The parameters of focal interest. If engine="l", the focal parameters should be specified using
the lavaan script. If engine="m", the focal parameters should be specified based on the Mplus output file design0.out.}

\item{complete.var}{Specify if there are any variables that need to have complete data collected across all the participating subjects}

\item{eval.budget}{Budget constraints. If the researcher wishes to search for designs under the budget limit, you to provide the remaining available budget that can be used for data collection}

\item{rm.budget}{The amount of remaining budget avaialbe for data collection}

\item{distal.var}{Any distal variables included in the model that would have complete data}

\item{seed}{Random seet for simulation}
}
\value{
An object containing the information of the optimal PM design, with highest power for testing the
focal parameters, compared with other PM designs
}
\description{
Function to search for the optimal planned missing design using lavaan/simsem
}
\seealso{
\code{\link{simPM}} which wraps this function
}
