% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced-simsem.R
\name{balance.miss.l}
\alias{balance.miss.l}
\title{Function to search for the optimal item-level PM design using lavaan/simsem}
\usage{
balance.miss.l(popModel, analyzeModel, NAMES, Time, Time.complete, k, pc,
  pd, costmx, n, nreps, focal.param, complete.var = NULL,
  eval.budget = T, rm.budget = NULL, distal.var = NULL,
  seed = 1234)
}
\arguments{
\item{popModel}{The data generation model (population model) specified using lavaan script}

\item{analyzeModel}{The analysis model,specified using lavaan script. The analysis model can be different from the population model.}

\item{NAMES}{A vector containing the names of the observed variables. The variable names must be ordered chronologically, by the time (wave) they are measured.}

\item{Time}{The total number of time points (or waves of data collection).}

\item{Time.complete}{Number of waves of data collection that have been completed before the funding cut occurs.}

\item{k}{The number of observed variables collected at each wave.}

\item{pc}{Proportion of completers. The proportion of subjects that will participate in all of the following waves of data collection and provide complete data. This must be greater than 0.}

\item{pd}{The proportion of subjects that will not participate in any of the following waves of data collection (i.e., drop from the longitudinal study). This can be 0.}

\item{costmx}{A vector containing the unit cost of each observed variable that is yet to be measured (post the funding cut). The cost is assumed to be constant across subjects, but it is allowed to vary across variables and across waves.}

\item{n}{The total sample size as initially planned.}

\item{nreps}{Number of replications for Monte Carlo simulations.}

\item{focal.param}{The parameters of focal interest. If engine="l", the focal parameters should be specified using
the lavaan script. If engine="m", the focal parameters should be specified based on the Mplus output file design0.out.}

\item{complete.var}{Specify the names of the variable(s) if there are any variable(s) that need to have complete data collected across all the participating subjects.}

\item{eval.budget}{Logical, indicating whether there is any budget constraint. If the user wishes to search for PM designs under the budget limit, they need to specify the amount of the remaining available budget that can be used for future data collection.}

\item{rm.budget}{The amount of remaining budget avaialbe for future data collection.}

\item{distal.var}{Specify the names of the variables, if there are any time-independent distal variables included in the model that are not subject planned missingness.}

\item{seed}{seed for random number generation.}
}
\value{
An object containing the information of the optimal item-level PM design, with highest power for testing the
focal parameters, compared with other candidate PM designs
}
\description{
Function to search for the optimal item-level PM design using lavaan/simsem
}
\seealso{
\code{\link{simPM}} which wraps this function
}
