% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-nm-1_simsem.R
\name{opt1.simsem}
\alias{opt1.simsem}
\title{Search for the optimal missing pattern with only one missing measured variable. An internal function for forward assembly.}
\usage{
opt1.simsem(popModel, analyzeModel, NAMES, distal.var, n, nreps, seed,
  Time, k, Time.complete, costmx, pc, pd, focal.param,
  complete.var = NULL)
}
\arguments{
\item{popModel}{The data generation model (population model) specified using lavaan script.}

\item{analyzeModel}{The analysis model, specified using lavaan script. The analysis model can be different from the population model.}

\item{NAMES}{A character vector containing the names of the observed
variables. The variable names must be ordered chronologically, by
the time (wave) they are measured.}

\item{distal.var}{Char vector. Specify the name(s) of the distal
variables. User needs to specify this argument if there are any 
time-independent distal variables included in the model that are not
subject to planned missingness.}

\item{n}{The total sample size as initially planned.}

\item{nreps}{Number of replications for Monte Carlo simulations.}

\item{seed}{The random seed for random number generation.}

\item{Time}{Numeric. The total number of time points (or total number
of waves of data collection).}

\item{k}{Numeric. The number of observed variables collected at each
wave.}

\item{Time.complete}{Numeric. Number of waves of data collection that
have been completed before the funding cut occurs.}

\item{costmx}{A numeric vector containing the unit cost of each
observed variable that is yet to be measured (post the funding cut).
The cost is assumed to be constant across subjects, but it is
allowed to vary across variables and across waves.}

\item{pc}{Numeric. Proportion of completers: the proportion of subjects
that will participate in all of the following waves of data
collection and provide complete data. This must be greater than 0.}

\item{pd}{Numeric. The proportion of subjects that will not participate
in any of the following waves of data collection (i.e., people who
will drop from the longitudinal study). This value can be 0.}

\item{focal.param}{Character vector. The parameters of focal interest.
The focal parameters should be specified in the format of 
\code{lavaan} script.}

\item{complete.var}{Char vector. Specify the name(s) of the 
variable(s) if there are any variable(s) that need to have complete
data collected from all the participating subjects.}
}
\value{
An object containing the information of the optimal missing
   data pattern containing only one missing observed variable. The
   optimal pattern is the one that yields highest statistical power for
   testing the focal parameters, compared to other patterns with only
   one missing observed variable.
}
\description{
\code{opt1.simsem} is an internal function that runs simulations using \code{lavaan} and \code{simsem}. It returns the optimal missing pattern that only contains one missing measured variable. This is the first step of forward assembly.
}
\seealso{
\code{\link{simPM}} which is a warpper function for this
   function.
}
