% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpattern.R
\name{plotPM}
\alias{plotPM}
\title{Plot the missing data patterns for the optimal PM design.}
\usage{
plotPM(object, colbr = "PRGn", col = c("antiquewhite1", "firebrick"),
  row.names = T, labels = T, fontsize_col = 20, fontsize_row = 14,
  fontsize = 14, angle_col = 45, legend = T, main = "", ...)
}
\arguments{
\item{object}{The simPM object.}

\item{colbr}{Specify the colors for different waves. Default is "PRGn".}

\item{col}{Specify the colors for complete vs. missing data. 
Default is c("antiquewhite1","firebrick").}

\item{labels}{logical scalar, indicating whether the label for waves is
needed. Default is TRUE.}

\item{fontsize_col}{Specify the font size for the column labels. Default
is 20.}

\item{fontsize_row}{Specify the font size for the row labels. Default is
14.}

\item{fontsize}{Specify the font size for the legend. Default is 14.}

\item{angle_col}{Specify the angle of how the column labels are
displayed. Default is 45.}

\item{legend}{Logical scalar, indicating whether the legend is shown.
Default is TRUE.}

\item{main}{Specify the plot title.}

\item{...}{Any additional arguments for \code{\link[pheatmap]{pheatmap}}.}
}
\description{
\code{plotPM} plots the optimal PM design's missing patterns as a
  heatmap.
}
\examples{
\dontrun{
plotPM(wave.out)
plotPM(indicator.out)
plotPM(forward.out,labels=F,col=c("gray96","gray35"),fontsize_row=26,fontsize=18,fontsize_col=26)
}
}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
