% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres-design.R
\name{pres.design}
\alias{pres.design}
\title{To check the simulation results for lower level designs in forward selection}
\usage{
pres.design(opt.results, max.mk)
}
\value{
An object containing the information of the optimal PM design, with highest power for testing the
focal parameters, compared with other PM designs
}
\description{
To check the simulation results for lower level designs in forward selection
}
\seealso{
\code{\link{simPM}} which wraps this function
}
