% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave-missing.R
\name{wave.miss}
\alias{wave.miss}
\title{Searching for wave-level PM designs (Mplus based)}
\usage{
wave.miss(VNAMES, distal.var = NULL, n, nreps, seed, Time, k,
  Time.complete, costmx, pc, pd, design0.out, focal.param,
  eval.budget = T, rm.budget = NULL, complete.wave = NULL)
}
\arguments{
\item{VNAMES}{A vector containing the names of the observed variables. The variable names must be ordered chronologically, by the time (wave) they are measured.}

\item{distal.var}{Specify the names of the variables, if there are any time-independent distal variables included in the model that are not subject planned missingness.}

\item{n}{The total sample size as initially planned.}

\item{nreps}{Number of replications for Monte Carlo simulations.}

\item{seed}{seed for random number generation.}

\item{Time}{The total number of time points (or waves of data collection).}

\item{k}{The number of observed variables collected at each wave.}

\item{Time.complete}{Number of waves of data collection that have been completed before the funding cut occurs.}

\item{costmx}{A vector containing the unit cost of each observed variable that is yet to be measured (post the funding cut). The cost is assumed to be constant across subjects, but it is allowed to vary across variables and across waves.}

\item{pc}{Proportion of completers. The proportion of subjects that will participate in all of the following waves of data collection and provide complete data. This must be greater than 0.}

\item{pd}{The proportion of subjects that will not participate in any of the following waves of data collection (i.e., drop from the longitudinal study). This can be 0.}

\item{design0.out}{Mplus output file which contains the a priori power analysis (i.e., simulations for sample size planning) results for this specific model assuming a complete data design. Theoretically, such analysis was supposed to be conducted before the study began.}

\item{focal.param}{The parameters of focal interest. The focal parameters should be specified based on the Mplus output file design0.out.}

\item{eval.budget}{Logical, indicating whether there is any budget constraint. If the user wishes to search for PM designs under the budget limit, they need to specify the amount of the remaining available budget that can be used for future data collection.}

\item{rm.budget}{The amount of remaining budget avaialbe for future data collection.}

\item{complete.wave}{Specify the wave(s) if there are any waves that need to have complete data collected across all the participants.}
}
\value{
An object containing the information of the optimal PM design, with highest power for testing the
focal parameters, compared with other PM designs
}
\description{
Searching for wave-level PM designs (Mplus based)
}
\seealso{
\code{\link{simPM}} which wraps this function
}
